+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format of the file is based on a template from [Keep a Changelog](http://kee
 
 ### Changed
 - DistanceTransform: integrated functionality to extract pre-cursor infos
+- DrawRegion2DSet: added parameters to specify desired size of result image directly
 - ParticleDetectorUWT2D: visibility of some internal methods updated
 
 ### Deprecated